fix cache race condition
Problem: Assume a page is uncached (by build-in phpFastCache page-cache). Then a request for the tag-index-page comes in. On this request the page gets cached with the meta `template` set to "tag". If the cached page is visited afterwards it is served with the wrong "tag" template.

Solution: move the setting of the template for the tag-page later in the cycle so that it's not cached as meta property.